RF-10128

+++ input/ui/src/main/java/org/richfaces/request/MultipartRequest25.java
@@ -28,6 +28,7 @@ import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
 
+import org.richfaces.exception.FileUploadException;
 import org.richfaces.log.Logger;
 import org.richfaces.log.RichfacesLogger;
 import org.richfaces.model.UploadedFile;
@@ -57,19 +58,37 @@ public class MultipartRequest25 extends BaseMultipartRequest {
         }
     };
 
-    private Multimap<String, String> params;
-    
-    private Iterable<UploadedFile> uploadedFiles;
-    
+    private MultipartRequestParser requestParser;
+
     private ResponseState responseState;
 
+    private Iterable<UploadedFile> uploadedFiles;
+
+    private Multimap<String, String> params;
+
     public MultipartRequest25(HttpServletRequest request, String uploadId, ProgressControl progressControl, 
-        Multimap<String, String> params, Iterable<UploadedFile> uploadedFiles, ResponseState responseState) {
+        MultipartRequestParser requestParser) {
+        
         super(request, uploadId, progressControl);
 
-        this.params = params;
-        this.uploadedFiles = uploadedFiles;
-        this.responseState = responseState;
+        this.requestParser = requestParser;
+    }
+
+    private void parseIfNecessary() {
+        if (responseState != null) {
+            return;
+        }
+        
+        try {
+            requestParser.parse();
+            
+            uploadedFiles = requestParser.getUploadedFiles();
+            params = requestParser.getParameters();
+            responseState = ResponseState.ok;
+        } catch (FileUploadException e) {
+            LOGGER.error(e.getMessage(), e);
+            responseState = ResponseState.serverError;
+        }
     }
     
     @SuppressWarnings({ "rawtypes", "unchecked" })
@@ -84,6 +103,7 @@ public class MultipartRequest25 extends BaseMultipartRequest {
             result.add(name);
         }
         
+        parseIfNecessary();
         result.addAll(params.keySet());
         
         return Iterators.asEnumeration(result.iterator());
@@ -91,11 +111,13 @@ public class MultipartRequest25 extends BaseMultipartRequest {
     
     @Override
     public String getParameter(String name) {
+
         String parameter = super.getParameter(name);
         if (parameter != null) {
             return parameter;
         }
         
+        parseIfNecessary();
         Collection<String> values = params.get(name);
         
         if (values.isEmpty()) {
@@ -112,6 +134,7 @@ public class MultipartRequest25 extends BaseMultipartRequest {
             return parameterValues;
         } 
         
+        parseIfNecessary();
         Collection<String> values = params.get(name);
         
         if (values.isEmpty()) {
@@ -125,28 +148,35 @@ public class MultipartRequest25 extends BaseMultipartRequest {
     @Override
     public Map getParameterMap() {
         Map parameterMap = Maps.newHashMap(super.getParameterMap());
+        parseIfNecessary();
         parameterMap.putAll(Maps.transformValues(params.asMap(), MULTIMAP_VALUE_TRANSFORMER));
         
         return parameterMap;
     }
     
     public Iterable<UploadedFile> getUploadedFiles() {
+        parseIfNecessary();
+        
         return uploadedFiles;
     }
     
     public void release() {
         super.release();
         
-        for (UploadedFile uploadedFile : uploadedFiles) {
-            try {
-                uploadedFile.delete();
-            } catch (IOException e) {
-                LOGGER.error(e.getMessage(), e);
+        if (uploadedFiles != null) {
+            for (UploadedFile uploadedFile : uploadedFiles) {
+                try {
+                    uploadedFile.delete();
+                } catch (IOException e) {
+                    LOGGER.error(e.getMessage(), e);
+                }
             }
         }
     }
     
     public ResponseState getResponseState() {
+        parseIfNecessary();
+        
         return responseState;
     }
